Outlier detection for observation for the log-Laplace model

This returns a unique number corresponding to the Bayes Factor associated to the test \(M_0: \Lambda_{obs} = \lambda_{ref}\) versus \(M_1: \Lambda_{obs} \neq \lambda_{ref}\) (with all other \(\Lambda_j, \neq obs\) free). The value of \(\lambda_{ref}\) is required as input. The user should expect long running times for the log-Student’s t model, in which case a reduced chain given \(\Lambda_{obs} = \lambda_{ref}\) needs to be generated

BF_lambda_obs_LLAP(obs, ref, X, chain)

Arguments

obs
Indicates the number of the observation under analysis

ref
Reference value \(\lambda_{ref}\) or \(u_{ref}\)

X
Design matrix with dimensions \(n\) x \(k\) where \(n\) is the number of observations and \(k\) is the number of covariates (including the intercept).

chain
MCMC chains generated by a BASSLINE MCMC function updates