Deviance information criterion for the log-exponential power model

Deviance information criterion is based on the deviance funcion \(D(\theta, y) = -2 log(f(y|\theta))\) but also incorporates a penalization factor of the complexity of the model

DIC_LEP(Time, Cens, X, chain, set = 1, eps_l = 0.5, eps_r = 0.5)

Arguments

Time
Vector containing the survival times.

Cens
Censoring indication (1: observed, 0: right-censored).

X
Design matrix with dimensions \(n\) x \(k\) where \(n\) is the number of observations and \(k\) is the number of covariates (including the intercept).

chain
MCMC chains generated by a BASSLINE MCMC function

set
Indicator for the use of set observations (1: set observations, 0: point observations).

eps_l
Lower imprecision \((\epsilon_l)\) for set observations (default value: 0.5).

eps_r
Upper imprecision \((\epsilon_r)\) for set observations (default value: 0.5)

Examples

library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations (especially for the log-exponential power model).

LEP <- MCMC_LEP(N = 1000, thin = 20, burn = 40, Time = cancer[,1],
                Cens = cancer[,2], X = cancer[,3:11])

#> Sampling initial betas from a Normal(0, 1) distribution 
#> Initial beta 1 : -0.63 
#>  Initial beta 2 : 0.66 
#>  Initial beta 3 : -0.29 
#>  Initial beta 4 : 0.56 
#>  Initial beta 5 : 1.86 
#>  Initial beta 6 : 0.66 
#>  Initial beta 7 : 1.5 
#>  Initial beta 8 : -1.19 
#>  Initial beta 9 : 1.06 
#> 
#> Sampling initial sigma^2 from a Gamma(2, 2) distribution 
#> Initial sigma^2 : 0.71 
#> 
#> Sampling initial alpha from a Uniform(1, 2) distribution 
#> Initial alpha : 1.03 
#> 
#> AR beta 1 : 0.57 
#>  AR beta 2 : 0.66 
#>  AR beta 3 : 0.73 
#>  AR beta 4 : 0.72 
#>  AR beta 5 : 0.8 
#>  AR beta 6 : 0.02 
#>  AR beta 7 : 0.07 
#>  AR beta 8 : 0.01 
#>  AR beta 9 : 0.74 
#> AR sigma2 : 0.91 
#> AR alpha : 0.09 

LEP.DIC <- DIC_LEP(Time = cancer[,1], Cens = cancer[,2], X = cancer[,3:11],
                   chain = LEP)

#> Effective number of parameters : 5.75 
#> Actual number of parameters : 11