VA Lung Cancer Trial Datatset

Data from a trial in which a therapy (standard or test chemotherapy) was randomly applied to 137 patients who were diagnosed with inoperable lung cancer. The survival times of the patients were measured in days since treatment.

cancer

Format

A matrix with 137 rows and 8 variables:

Time

Survival time (in days)

Cens

0 or 1. If 0 the observation is right censored

Intercept

The intercept

Treat

The treatment applied to the patient (0: standard, 1: test)

Type.1

The histological type of the tumor (1: type 1, 0: otherwise)

Type.2

The histological type of the tumor (1: type 2, 0: otherwise)

Type.3

The histological type of the tumor (1: type 3, 0: otherwise)

Status

A continuous index representing the status of the patient: 10—30 completely hospitalized, 40—60 partial confinement, 70—90 able to care for self.

MFD

The time between the diagnosis and the treatment (in months)

Age

Age (in years)

Prior

Prior therapy, 0 or 10

Source

Appendix I of Kalbfleisch and Prentice (1980).